pyproject.toml

[tool.isort]
profile = "black"
py_version = 310
color_output = false

[tool.flake8]
max-line-length = 88

[tool.black]
line-length = 88
target-version = ['py310']
exclude = 'migrations'

[tool.poetry]
name = "planscape"
version = "0.1.0"
package-mode = false
description = "planscape"
authors = ["SIG-GIS"]

[tool.poetry.dependencies]
python = "^3.10"
dj-rest-auth = ">=5.0.2,<5.1.0"
django = "^4.2.11"
django-allauth = ">=0.57.0,<0.58.0"
django-cors-headers = "^4.3.1"
django-filter = ">=21.1.0,<21.2.0"
djangorestframework = ">=3.14.0,<3.15.1"
djangorestframework-gis = ">=0.18.0"
djangorestframework-simplejwt = ">=5.2.2,<5.3.0"
numpy = "^1.26.4"
pandas = "^2.2.2"
psycopg2-binary = ">=2.9.3"
python-decouple = "3.6"
rasterio = "^1.3.10"
rasterstats = "^0.19.0"
typing-extensions = "^4.11.0"
uwsgi = "^2.0.25.1"
boto3 = "^1.34.88"
pytz = "^2024.1"
sentry-sdk = "^1.45.0"
django-password-policies-validator = ">=1.0.2"
black = "23.7.0"
humanize = "4.8.0"
shapely = "2.0.2"
fiona = "1.9.5"
celery = "5.3.6"
redis = "5.0.1"
kombu = "5.3.4"
shortuuid = "1.0.11"
jsonschema = ">=4.21,<5.0"
gunicorn = "^22.0.0"
pyopenssl = "24.1.0"
django-treebeard = "^4.7.1"


[tool.poetry.dev-dependencies]
pytest = "7.1.2"
pytest-cov = "^3.0.0"
pytest-django = "^4.5.2"
pytest-mock = "^3.8.2"
jwcrypto = "^1.3.1"
responses = "^0.21.0"
moto = "^4.0.0"
black = "23.7.0"
isort = "^5.12.0"
memory-profiler = "^0.61.0"
django-extensions = "3.2.3"
pytest-watch = "^4.2.0"
pre-commit = "^2.20.0"
flake8 = "^6.0.0"
coverage = "^6.4.4"
freezegun = "^1.2.2"
requests-mock = "^1.9.3"
mypy = "^0.971"
types-python-dateutil = "^2.8.18"
pytest-dotenv = "^0.5.2"
types-cachetools = "^5.2.1"
typed-ast = "^1.5.4"

[tool.poetry.group.dev.dependencies]
colorama = "^0.4.6"
poetry-plugin-export = "^1.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--cov --cov-append --cov-report term-missing --envfile .test.env --no-cov-on-fail --no-header --disable-warnings -rf --durations=5 --no-migrations"
DJANGO_SETTINGS_MODULE = "planscape.settings"
python_files = "tests.py test_*.py *_tests.py"
